bin/generate_corpus/requirements.txt
aspose-words
requests
